requirements.txt
amqp==5.2.0
anyio==4.3.0
billiard==4.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fauna==1.2.0b4
Flask==3.0.2
future==1.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.6
MarkupSafe==2.1.5
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
